Small code clean-ups.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1531 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/LogitScaleEngine.cpp
@@ -93,7 +93,7 @@ QwtScaleDiv LogitScaleEngine::divideScale(double x1, double x2,
     QwtScaleDiv scaleDiv;
     if ( stepSize != 0.0 ){
         QwtValueList ticks[QwtScaleDiv::NTickTypes];
-		buildTicks(interval, stepSize, ticks);
+		buildTicks(interval, (int)stepSize, ticks);
         scaleDiv = QwtScaleDiv(interval, ticks);
     }
