+++ qtiplot/src/plot2D/ProbabilityScaleEngine.cpp
@@ -94,7 +94,7 @@ QwtScaleDiv ProbabilityScaleEngine::divideScale(double x1, double x2,
     QwtScaleDiv scaleDiv;
     if ( stepSize != 0.0 ){
         QwtValueList ticks[QwtScaleDiv::NTickTypes];
-		buildTicks(interval, stepSize, ticks);
+		buildTicks(interval, (int)stepSize, ticks);
         scaleDiv = QwtScaleDiv(interval, ticks);
     }
